# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: GNNcab - Multi-city ride dispatch and driver tracking platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trips:
    get:
      operationId: getTrips
      tags: [trips]
      summary: Get trips (filtered by role)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: driverId
          in: query
          schema:
            type: string
        - name: customerId
          in: query
          schema:
            type: string
        - name: cityId
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: number
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated list of trips
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripsListResponse"
    post:
      operationId: createTrip
      tags: [trips]
      summary: Request a new trip
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTripRequest"
      responses:
        "201":
          description: Trip created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trips/{tripId}:
    get:
      operationId: getTrip
      tags: [trips]
      summary: Get trip by ID
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Trip details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripDetails"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trips/{tripId}/status:
    put:
      operationId: updateTripStatus
      tags: [trips]
      summary: Update trip status
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTripStatusRequest"
      responses:
        "200":
          description: Trip status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "400":
          description: Invalid transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trips/{tripId}/events:
    get:
      operationId: getTripEvents
      tags: [trips]
      summary: Get events for a trip
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Trip events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TripEvent"
  /trips/{tripId}/cancel:
    post:
      operationId: cancelTrip
      tags: [trips]
      summary: Cancel a trip
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelTripRequest"
      responses:
        "200":
          description: Trip cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
...
  /driver/trips:
    get:
      operationId: getDriverTrips
      tags: [trips]
      summary: Get trips for current driver
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Driver trips
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripsListResponse"
...
  /customer/trips:
    get:
      operationId: getCustomerTrips
      tags: [trips]
      summary: Get trips for current customer
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Customer trips
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripsListResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
      required:
        - error
        - message
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
          nullable: true
        role:
          type: string
          enum: [admin, driver, customer]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - firstName
        - lastName
        - role
        - isActive
        - createdAt
        - updatedAt
...
    DriverDetails:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
          nullable: true
        licenseNumber:
          type: string
          nullable: true
        vehicleMake:
          type: string
          nullable: true
        vehicleModel:
          type: string
          nullable: true
        vehicleYear:
          type: number
          nullable: true
        vehicleColor:
          type: string
          nullable: true
        vehiclePlate:
          type: string
          nullable: true
        status:
          type: string
          enum: [AVAILABLE, BUSY, OFFLINE]
        isVerified:
          type: boolean
        currentLat:
          type: number
          nullable: true
        currentLng:
          type: number
          nullable: true
        lastSeen:
          type: string
          format: date-time
          nullable: true
        cityId:
          type: string
          nullable: true
        rating:
          type: number
          nullable: true
        totalTrips:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - email
        - firstName
        - lastName
        - status
        - isVerified
        - totalTrips
        - createdAt
        - updatedAt
...
    Trip:
      type: object
      properties:
        id:
          type: string
        customerId:
          type: string
        driverId:
          type: string
          nullable: true
        cityId:
          type: string
          nullable: true
        pickupLat:
          type: number
        pickupLng:
          type: number
        pickupAddress:
          type: string
        dropoffLat:
          type: number
        dropoffLng:
          type: number
        dropoffAddress:
          type: string
        status:
          type: string
          enum: [REQUESTED, DRIVER_ASSIGNED, DRIVER_ACCEPTED, DRIVER_REJECTED, DRIVER_EN_ROUTE, DRIVER_ARRIVED, TRIP_STARTED, TRIP_COMPLETED, TRIP_CANCELLED, DRIVER_NO_RESPONSE, TIMEOUT]
        distanceKm:
          type: number
          nullable: true
        estimatedFare:
          type: number
          nullable: true
        finalFare:
          type: number
          nullable: true
        currency:
          type: string
        paymentMethod:
          type: string
          enum: [cash, card, online]
        cancellationReason:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        requestedAt:
          type: string
          format: date-time
        assignedAt:
          type: string
          format: date-time
          nullable: true
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - customerId
        - pickupLat
        - pickupLng
        - pickupAddress
        - dropoffLat
        - dropoffLng
        - dropoffAddress
        - status
        - currency
        - paymentMethod
        - requestedAt
        - createdAt
        - updatedAt
    TripDetails:
      type: object
      properties:
        id:
          type: string
        customerId:
          type: string
        driverId:
          type: string
          nullable: true
        cityId:
          type: string
          nullable: true
        pickupLat:
          type: number
        pickupLng:
          type: number
        pickupAddress:
          type: string
        dropoffLat:
          type: number
        dropoffLng:
          type: number
        dropoffAddress:
          type: string
        status:
          type: string
          enum: [REQUESTED, DRIVER_ASSIGNED, DRIVER_ACCEPTED, DRIVER_REJECTED, DRIVER_EN_ROUTE, DRIVER_ARRIVED, TRIP_STARTED, TRIP_COMPLETED, TRIP_CANCELLED, DRIVER_NO_RESPONSE, TIMEOUT]
        distanceKm:
          type: number
          nullable: true
        estimatedFare:
          type: number
          nullable: true
        finalFare:
          type: number
          nullable: true
        currency:
          type: string
        paymentMethod:
          type: string
          enum: [cash, card, online]
        cancellationReason:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        requestedAt:
          type: string
          format: date-time
        assignedAt:
          type: string
          format: date-time
          nullable: true
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        customer:
          $ref: "#/components/schemas/UserProfile"
        driver:
          $ref: "#/components/schemas/DriverDetails"
          nullable: true
        events:
          type: array
          items:
            $ref: "#/components/schemas/TripEvent"
        payment:
          $ref: "#/components/schemas/Payment"
          nullable: true
      required:
        - id
        - customerId
        - pickupLat
        - pickupLng
        - pickupAddress
        - dropoffLat
        - dropoffLng
        - dropoffAddress
        - status
        - currency
        - paymentMethod
        - requestedAt
        - createdAt
        - updatedAt
        - customer
        - events
    TripEvent:
      type: object
      properties:
        id:
          type: string
        tripId:
          type: string
        status:
          type: string
        actorId:
          type: string
          nullable: true
        actorRole:
          type: string
          nullable: true
        metadata:
          type: object
          additionalProperties: true
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - tripId
        - status
        - createdAt
    CreateTripRequest:
      type: object
      properties:
        pickupLat:
          type: number
        pickupLng:
          type: number
        pickupAddress:
          type: string
        dropoffLat:
          type: number
        dropoffLng:
          type: number
        dropoffAddress:
          type: string
        paymentMethod:
          type: string
          enum: [cash, card, online]
        notes:
          type: string
        cityId:
          type: string
      required:
        - pickupLat
        - pickupLng
        - pickupAddress
        - dropoffLat
        - dropoffLng
        - dropoffAddress
        - paymentMethod
    UpdateTripStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [DRIVER_ACCEPTED, DRIVER_REJECTED, DRIVER_EN_ROUTE, DRIVER_ARRIVED, TRIP_STARTED, TRIP_COMPLETED]
        metadata:
          type: object
          additionalProperties: true
      required:
        - status
    CancelTripRequest:
      type: object
      properties:
        reason:
          type: string
      required:
        - reason
    TripsListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Trip"
        total:
          type: number
        page:
          type: number
        limit:
          type: number
        totalPages:
          type: number
      required:
        - data
        - total
        - page
        - limit
        - totalPages
    Payment:
      type: object
      properties:
        id:
          type: string
        tripId:
          type: string
        amount:
          type: number
        currency:
          type: string
        method:
          type: string
          enum: [cash, card, online]
        status:
          type: string
          enum: [PENDING, COMPLETED, FAILED, REFUNDED]
        transactionId:
          type: string
          nullable: true
        metadata:
          type: object
          additionalProperties: true
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - tripId
        - amount
        - currency
        - method
        - status
        - createdAt
        - updatedAt
...